% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact.R
\name{fact}
\alias{fact}
\title{A Factorial Function}
\usage{
fact(x)
}
\arguments{
\item{x}{Positive integer (please)}
}
\description{
This function uses recursion to compute the factorial of a positive integer.
}
\examples{
fact(3) = 6
}
\keyword{inefficiency}
